requirements.txt
spotipy
flask
python-dotenv